# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: STEEP/PESTLE Foresight Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/pick-name:
    post:
      operationId: pickName
      tags: [auth]
      summary: Choose a roster identity (or self-add)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PickNameBody" }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "400":
          description: Invalid
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  /auth/instructor-login:
    post:
      operationId: instructorLogin
      tags: [auth]
      summary: Sign in as instructor with passcode
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstructorLoginBody" }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Collection" }
    post:
      operationId: createCollection
      tags: [collections]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCollectionBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
    patch:
      operationId: updateCollection
      tags: [collections]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCollectionBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
  /collections/{id}/items:
    post:
      operationId: addCollectionItem
      tags: [collections]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddCollectionItemBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
  /collections/{id}/items/{itemId}:
    delete:
      operationId: removeCollectionItem
      tags: [collections]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: itemId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Removed }

  # ---------- Scenarios ----------
...
  /comments/{entityId}:
    get:
      operationId: listComments
      tags: [comments]
      parameters:
        - { name: entityId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Comment" }
    post:
      operationId: addComment
      tags: [comments]
      parameters:
        - { name: entityId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddCommentBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
  /comments/{entityId}/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      parameters:
        - { name: entityId, in: path, required: true, schema: { type: string } }
        - { name: commentId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }

  # ---------- Analytics ----------
...
  /analytics/student/{id}:
    get:
      operationId: getStudentProfile
      tags: [analytics]
      summary: Full activity profile for a student
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudentProfile" }
  /analytics/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [analytics]
      summary: Recent participation events across the class
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActivityEvent" }

  # ---------- Export ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]

    # ----- Session / roster -----
    SessionUser:
      type: object
      properties:
        id: { type: ["string", "null"] }
        name: { type: ["string", "null"] }
        role:
          type: ["string", "null"]
          enum: ["student", "instructor", null]
        signedIn: { type: boolean }
      required: [signedIn]
    PickNameBody:
      type: object
      properties:
        rosterId: { type: ["string", "null"] }
        newName: { type: ["string", "null"] }
      required: []
    InstructorLoginBody:
      type: object
      properties:
        passcode: { type: string }
      required: [passcode]
    RosterMember:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        role:
          type: string
          enum: [student, instructor]
        createdAt: { type: string, format: date-time }
      required: [id, name, role, createdAt]
...
    Source:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        url: { type: string }
        publication: { type: ["string", "null"] }
        author: { type: ["string", "null"] }
        publishedDate: { type: ["string", "null"] }
        type:
          type: string
          enum: [news, research, report, blog, video, podcast, social, dataset, other]
        credibilityNote: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, title, url, type, createdAt]
...
    Signal:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        summary: { type: string }
        category: { type: string }
        novelty: { type: string }
        timeHorizon: { type: string }
        keywords:
          type: array
          items: { type: string }
        implications: { type: string }
        context: { type: string }
        assumptions: { type: string }
        unknowns: { type: string }
        ethicsEquity: { type: string }
        leadingIndicators: { type: string }
        strategicResponse: { type: string }
        impactRating: { type: integer }
        uncertaintyRating: { type: integer }
        plausibilityRating: { type: integer }
        status:
          type: string
          enum: [draft, submitted, needs_revision, archived]
        sources:
          type: array
          items: { $ref: "#/components/schemas/Source" }
        authorId: { type: string }
        authorName: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        submittedAt: { type: ["string", "null"], format: date-time }
      required: [id, title, summary, category, novelty, timeHorizon, keywords, implications, context, assumptions, unknowns, ethicsEquity, leadingIndicators, strategicResponse, impactRating, uncertaintyRating, plausibilityRating, status, sources, authorId, authorName, createdAt, updatedAt]
...
    Trend:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        claim: { type: string }
        rationale: { type: string }
        category: { type: string }
        timeHorizon: { type: string }
        confidenceRating: { type: integer }
        impactRating: { type: integer }
        status:
          type: string
          enum: [preliminary, established, contested]
        supportingSignalIds:
          type: array
          items: { type: string }
        contradictingSignalIds:
          type: array
          items: { type: string }
        authorId: { type: string }
        authorName: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, claim, rationale, category, timeHorizon, confidenceRating, impactRating, status, supportingSignalIds, contradictingSignalIds, authorId, authorName, createdAt, updatedAt]
...
    CollectionItem:
      type: object
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [signal, trend]
        refId: { type: string }
        note: { type: string }
        addedBy: { type: string }
        addedByName: { type: string }
        addedAt: { type: string, format: date-time }
      required: [id, kind, refId, note, addedBy, addedByName, addedAt]
    Collection:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        visibility:
          type: string
          enum: [private, class]
        items:
          type: array
          items: { $ref: "#/components/schemas/CollectionItem" }
        ownerId: { type: string }
        ownerName: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, description, visibility, items, ownerId, ownerName, createdAt, updatedAt]
    CreateCollectionBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        visibility:
          type: string
          enum: [private, class]
      required: [title, visibility]
    UpdateCollectionBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        visibility:
          type: string
          enum: [private, class]
    AddCollectionItemBody:
      type: object
      properties:
        kind:
          type: string
          enum: [signal, trend]
        refId: { type: string }
        note: { type: string }
      required: [kind, refId]

    # ----- Scenario -----
    ProblemStatement:
      type: object
      properties:
        id: { type: string }
        text: { type: string }
        signalIds:
          type: array
          items: { type: string }
      required: [id, text, signalIds]
    Persona:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        background: { type: string }
        needs: { type: string }
        problemStatementId: { type: string }
      required: [id, name, role, background, needs, problemStatementId]
    Journey:
      type: object
      properties:
        id: { type: string }
        personaId: { type: string }
        steps:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              detail: { type: string }
              signalIds:
                type: array
                items: { type: string }
            required: [label, detail, signalIds]
      required: [id, personaId, steps]
    SignalAnnotation:
      type: object
      properties:
        signalId: { type: string }
        role:
          type: string
          enum: [key_driver, uncertainty, wildcard, supporting]
      required: [signalId, role]
    Scenario:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        timeHorizon: { type: string }
        archetype: { type: string }
        environmentDescription: { type: string }
        vignette: { type: string }
        signalIds:
          type: array
          items: { type: string }
        signalAnnotations:
          type: array
          items: { $ref: "#/components/schemas/SignalAnnotation" }
        problemStatements:
          type: array
          items: { $ref: "#/components/schemas/ProblemStatement" }
        personas:
          type: array
          items: { $ref: "#/components/schemas/Persona" }
        journeys:
          type: array
          items: { $ref: "#/components/schemas/Journey" }
        currentStep: { type: integer }
        status:
          type: string
          enum: [draft, submitted]
        collectionId: { type: ["string", "null"] }
        authorId: { type: string }
        authorName: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        submittedAt: { type: ["string", "null"], format: date-time }
      required: [id, title, timeHorizon, archetype, environmentDescription, vignette, signalIds, signalAnnotations, problemStatements, personas, journeys, currentStep, status, authorId, authorName, createdAt, updatedAt]
...
    Comment:
      type: object
      properties:
        id: { type: string }
        entityId: { type: string }
        body: { type: string }
        authorId: { type: string }
        authorName: { type: string }
        authorRole:
          type: string
          enum: [student, instructor]
        createdAt: { type: string, format: date-time }
      required: [id, entityId, body, authorId, authorName, authorRole, createdAt]
    AddCommentBody:
      type: object
      properties:
        body: { type: string }
      required: [body]

    # ----- Analytics -----
    StudentParticipation:
      type: object
      properties:
        studentId: { type: string }
        studentName: { type: string }
        signalsTotal: { type: integer }
        signalsSubmitted: { type: integer }
        signalsDraft: { type: integer }
        trendsCount: { type: integer }
        collectionsCount: { type: integer }
        scenariosCount: { type: integer }
        scenariosSubmitted: { type: integer }
        commentsCount: { type: integer }
        lastActiveAt: { type: ["string", "null"], format: date-time }
        atRisk: { type: boolean }
      required: [studentId, studentName, signalsTotal, signalsSubmitted, signalsDraft, trendsCount, collectionsCount, scenariosCount, scenariosSubmitted, commentsCount, atRisk]
...
    StudentProfile:
      type: object
      properties:
        student: { $ref: "#/components/schemas/RosterMember" }
        participation: { $ref: "#/components/schemas/StudentParticipation" }
        signals:
          type: array
          items: { $ref: "#/components/schemas/Signal" }
        trends:
          type: array
          items: { $ref: "#/components/schemas/Trend" }
        collections:
          type: array
          items: { $ref: "#/components/schemas/Collection" }
        scenarios:
          type: array
          items: { $ref: "#/components/schemas/Scenario" }
        timeline:
          type: array
          items: { $ref: "#/components/schemas/ActivityEvent" }
      required: [student, participation, signals, trends, collections, scenarios, timeline]
    ActivityEvent:
      type: object
      properties:
        id: { type: string }
        actorId: { type: string }
        actorName: { type: string }
        type: { type: string }
        entityKind: { type: string }
        entityId: { type: string }
        entityTitle: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, actorId, actorName, type, entityKind, entityId, createdAt]