# lib/api-zod/src/generated/types/rosterMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * STEEP/PESTLE Foresight Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { RosterMemberRole } from "./rosterMemberRole";

export interface RosterMember {
  id: string;
  name: string;
  role: RosterMemberRole;
  createdAt: Date;
}